requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
Django==4.2.19
django-environ==0.12.0
djangorestframework==3.15.2
drf-yasg==1.21.8
exceptiongroup==1.2.2
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
jiter==0.8.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
openai==1.63.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.16
zipp==3.21.0
pytest
pytest-django